+++ QcloudApi/qcloudapi.py
@@ -128,6 +128,9 @@ class QcloudApi(object):
         elif module == 'athena':
             from .modules.athena import Athena
             service = Athena(config)
+        elif module == 'emr':
+            from .modules.emr import Emr
+            service = Emr(config)
         else:
             raise ValueError('module not exists')